chore(zsh): added stop command to the simple web server

+++ modules/home/zsh.nix
@@ -33,6 +33,12 @@
 
       function web {
         docker stop simple-web-server >/dev/null 2>&1
+
+        if [[ $1 == 'stop' ]]; then
+          echo "Stopped webserver"
+          return
+        fi
+
         CONTAINER=$(docker run -d --rm -ti --volume "`pwd`:/usr/share/nginx/html:ro" --name simple-web-server -p 127.0.0.1:5500:80 nginx)
         docker exec -tid ''${CONTAINER} bash -c "sleep 3600 && kill 1"
         echo "Started webserver running at: http://localhost:5500 (for 1 hour)"
